added camera null functionality

+++ server.js
@@ -35,23 +35,23 @@ http.listen(port, function() {
 });
 
 function connected(cameraInfo, socket){
-  var dir_path = dir + cameraInfo.camera + '/',
+  var sliced = [];
+  if(cameraInfo.camera) {
+    var dir_path = dir + cameraInfo.camera + '/',
     files = fs.readdirSync(dir_path), 
-    sliced = [],
     filterTime = new Date(new Date().getTime() - (config.stream.filterTime * 1000)).getTime();
-  if(files.length > 0) {
-    var after_dlt = _.each(files, function(file) {
-      if(fs.statSync(dir_path + file).ctime.getTime() < filterTime) {
-        fs.unlink(dir_path + file);
-      } 
-    });
-    if(after_dlt.length >=3) {
-      sliced = after_dlt.slice(after_dlt.length-3, after_dlt.length);
+    if(files.length > 0) {
+      var after_dlt = _.each(files, function(file) {
+        if(fs.statSync(dir_path + file).ctime.getTime() < filterTime) {
+          fs.unlink(dir_path + file);
+        } 
+      });
+      if(after_dlt.length >=3) {
+        sliced = after_dlt.slice(after_dlt.length-3, after_dlt.length);
+      }
     }
-    socket.emit('videoSend', {'videos' : sliced});
-  } else {
-    socket.emit('videoSend', {'videos' : sliced});
-  }  
+  }
+  socket.emit('videoSend', {'videos' : sliced});
 };
 
 watcher.on('ready', function() {
